to do

+++ src/graphql/furniture/furniture.resolvers.js
@@ -1,6 +1,9 @@
 import GraphQLUpload from 'graphql-upload/GraphQLUpload.mjs';
 import FurnitureService from '../../services/FurnitureService.js';
-import uploadFileToGCS from '../../utils/middleware/uploadFile.js';
+import {
+  uploadFileToGCS,
+  deleteFileFromGCS
+} from '../../utils/middleware/uploadFile.js';
 
 const resolvers = {
   Upload: GraphQLUpload,
@@ -10,32 +13,41 @@ const resolvers = {
     furniture: async (_, { id }) => await FurnitureService.getFurnitureById(id)
   },
   Mutation: {
-    createFurniture: async (_parent, { name, icon, category }, user, _info) => {
-      console.log('category:', category);
-      console.log('icon:', icon);
-      console.log('name:', name);
-      console.log('user:', user);
-
+    createFurniture: async (_parent, { name, icon, category }, user) => {
       if (!user) throw new Error('Unauthorized');
+
       if (user.role === 'admin') {
-        let iconUrl = null;
-
         if (icon) {
-          const { createReadStream, filename } = await icon;
-          const fileUrl = await uploadFileToGCS({ createReadStream, filename });
-          iconUrl = fileUrl;
-        }
+          try {
+            const { createReadStream } = await icon;
+            const updatedFilename = name;
+            const folder = 'icons';
 
-        return await FurnitureService.createFurniture({
-          name,
-          icon: iconUrl,
-          category
-        });
+            const fileUrl = await uploadFileToGCS({
+              createReadStream,
+              folder,
+              filename: updatedFilename
+            });
+
+            try {
+              const furniture = await FurnitureService.createFurniture({
+                name,
+                icon: fileUrl,
+                category
+              });
+              return furniture;
+            } catch (dbError) {
+              await deleteFileFromGCS({ folder, filename: updatedFilename });
+              throw dbError;
+            }
+          } catch (uploadError) {
+            throw new Error(`File upload failed: ${uploadError.message}`);
+          }
+        }
       }
       throw new Error('You must be an admin to create furniture');
     },
-
-    updateFurniture: async (_parent, { id, input }, user, _info) => {
+    updateFurniture: async (_parent, { id, input }, user) => {
       if (!user) throw new Error('Unauthorized');
       if (user.role === 'admin') {
         let iconUrl = null;
@@ -53,10 +65,22 @@ const resolvers = {
       }
       throw new Error('You must be an admin to update furniture');
     },
-    deleteFurniture: async (_parent, _args, user, _info) => {
+    deleteFurniture: async (_parent, _args, user) => {
       if (!user) throw new Error('Unauthorized');
+
       if (user.role === 'admin') {
-        return await FurnitureService.deleteFurniture(_args.id);
+        const furniture = await FurnitureService.getFurnitureById(_args.id);
+        if (furniture) {
+          const folder = 'icons';
+          const filename = furniture.name;
+
+          const [deleteFileResult, deleteFurnitureResult] = await Promise.all([
+            deleteFileFromGCS({ folder, filename }),
+            FurnitureService.deleteFurniture(_args.id)
+          ]);
+
+          return deleteFurnitureResult;
+        }
       }
     }
   }
